Move custom tab bar files to its own directory
